# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/api/inner/ws-gateway-service/client-publish": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "推送消息-内部"
                ],
                "summary": "推送链接消息",
                "parameters": [
                    {
                        "description": "推送消息",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/response.ClientPublishData"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.StandardResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/wsInner.publishResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/api/inner/ws-gateway-service/group-publish": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "推送消息-内部"
                ],
                "summary": "推送组消息",
                "parameters": [
                    {
                        "description": "推送消息",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/response.GroupPublishData"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.StandardResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/wsInner.publishResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/api/inner/ws-gateway-service/user-publish": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "推送消息-内部"
                ],
                "summary": "推送用户消息",
                "parameters": [
                    {
                        "description": "推送消息",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/response.UserPublishData"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.StandardResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/wsInner.publishResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/api/public/ws-gateway-service/group/msg/index": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "组消息-外部"
                ],
                "summary": "组消息",
                "parameters": [
                    {
                        "type": "string",
                        "name": "group_id",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "name": "index",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "name": "user_id",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.StandardResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/response.GroupPublishData"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/api/public/ws-gateway-service/user/msg/index": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "组消息-外部"
                ],
                "summary": "组消息",
                "parameters": [
                    {
                        "type": "string",
                        "name": "group_id",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "name": "index",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "name": "user_id",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.StandardResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/response.UserPublishData"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "response.ClientPublishData": {
            "type": "object",
            "properties": {
                "client_id": {
                    "description": "客户端id",
                    "type": "string"
                },
                "index": {
                    "description": "消息索引 提交消息时自增",
                    "type": "integer"
                },
                "message": {
                    "description": "消息数据 json字符串",
                    "type": "string"
                }
            }
        },
        "response.GroupPublishData": {
            "type": "object",
            "properties": {
                "group_id": {
                    "description": "消息组",
                    "type": "string"
                },
                "index": {
                    "description": "消息索引 提交消息时自增",
                    "type": "integer"
                },
                "message": {
                    "description": "消息数据 json字符串",
                    "type": "string"
                }
            }
        },
        "response.StandardResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "0 成功 1失败",
                    "type": "integer"
                },
                "data": {},
                "message": {
                    "description": "提示语",
                    "type": "string"
                }
            }
        },
        "response.UserPublishData": {
            "type": "object",
            "properties": {
                "index": {
                    "description": "消息索引 提交消息时自增",
                    "type": "integer"
                },
                "message": {
                    "description": "消息数据 json字符串",
                    "type": "string"
                },
                "user_id": {
                    "description": "用户id",
                    "type": "string"
                }
            }
        },
        "wsInner.publishResponse": {
            "type": "object",
            "properties": {
                "index": {
                    "description": "消息索引",
                    "type": "integer"
                }
            }
        }
    }
}